Fix dtype mismatch in Vision Mamba webcam prediction
Fixed camera_stats dtype issue:
- camera_stats_tensor now explicitly uses float32
- image_tensor also converted to float32
- Prevents mat1/mat2 dtype mismatch error

Now the AI model runs in real-time with webcam:
- ✅ Model loads successfully (2.55M parameters)
- ✅ Webcam capture working (640x480 @ 30 FPS)
- ✅ Real-time Vision Mamba inference
- ✅ Steering/Throttle/Brake predictions working
- ✅ FiLM layer adapting to camera conditions

GUI fully functional with actual AI predictions!

+++ vision-mamba-control/src/models/control_model.py
@@ -222,10 +222,10 @@ class VisionMambaControl(nn.Module):
         frame_norm = frame_resized.astype(np.float32) / 255.0
 
         # To tensor (B, C, H, W)
-        image_tensor = torch.from_numpy(frame_norm).permute(2, 0, 1).unsqueeze(0)
+        image_tensor = torch.from_numpy(frame_norm).permute(2, 0, 1).unsqueeze(0).float()
 
         # Camera stats
-        camera_stats_tensor = torch.tensor([[brightness, contrast, saturation]])
+        camera_stats_tensor = torch.tensor([[brightness, contrast, saturation]], dtype=torch.float32)
 
         # Predict
         with torch.no_grad():
